Extract PR tool registrations into pr-tools.ts (#tech-debt)
Move create_pr and get_pr_status MCP tool registrations from the
monolithic server.ts into a dedicated pr-tools.ts module, following
the same pattern as job-tools.ts and persona-tools.ts.
